+++ app/Config/Routes.php
@@ -6,4 +6,5 @@ use CodeIgniter\Router\RouteCollection;
  * @var RouteCollection $routes
  */
 $routes->get('/', 'Home::index');
-$routes->get('/users/register', 'Register::register');
+$routes->get('/users/register', 'RegisterController::register');
+$routes->get('/users/login', 'LoginController::login');
